Fix Export of EPS/PDF When Using Black yyaxis
If the yyaxis command is used to set the axes to be black, then we hit the issue of black axes being the same colour as the temporary black background used in export of EPS or PDF files.

The current method of fixing this only works for the last selected yaxis, as setting 'YColor' only affects that axis not both. As a result the figure crops incorrectly.

To work around this, once each of the black axes has been found, we can check these for the corresponding '?Axis' property to see if there is more than one rule. If so, we set each non-black ruler to be not quite black, returning its handle so it can be later set back to black.

As a result the cropping can now properly detect both axes.

This potentially fixes https://github.com/altmany/export_fig/issues/306.